packages/gateway/pyproject.toml
[project]
name = "a2ui-gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.10.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "langchain>=1.2.7",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "sse-starlette>=3.2.0",
    "uvicorn>=0.40.0",
]
